Migrate OTR room action
