+++ apps/meteor/app/otr/client/index.ts
@@ -1,4 +1,3 @@
 import './OTRRoom';
 import './OTR';
-import './tabBar';
 import './messageTypes';
